+++ client/src/components/Dashboard.js
@@ -1,7 +1,6 @@
 import React, {useState} from 'react';
 import { Link, withRouter, } from 'react-router-dom'
 import TasksDue from './TasksDue'
-import TaskClass from './TaskClass'
 import ListContainer from './ListContainer';
 
 const Dashboard = (props) => {
@@ -32,8 +31,7 @@ const Dashboard = (props) => {
   return (
     <>
       <Link to='dashboard'>Board</Link>
-      {/* <TasksDue /> */}
-      <TaskClass />
+      <TasksDue />
     </>
   )
 }
